add real tests and improve some implementations

+++ src/process.cpp
@@ -1,13 +1,25 @@
 #include <csignal>
 #include <cstdlib>
 #include <librdb/error.hpp>
+#include <librdb/pipe.hpp>
 #include <librdb/process.hpp>
 #include <memory>
 #include <sys/ptrace.h>
 #include <sys/wait.h>
 #include <unistd.h>
 
-std::unique_ptr<rdb::process> rdb::process::launch(std::filesystem::path path) {
+namespace {
+void exit_with_perror(rdb::pipe &channel, std::string const &prefix) {
+  auto message = prefix + ": " + std::strerror(errno);
+  channel.write(reinterpret_cast<std::byte *>(message.data()), message.size());
+  exit(-1);
+}
+} // namespace
+
+std::unique_ptr<rdb::process> rdb::process::launch(std::filesystem::path path,
+                                                   bool debug) {
+
+  pipe channel(/*close_on_exec=*/true);
 
   pid_t pid = 0;
 
@@ -18,16 +30,31 @@ std::unique_ptr<rdb::process> rdb::process::launch(std::filesystem::path path) {
   if (pid == 0) {
     // in child process
     // execute debugee
-    if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
-      error::send_errno("Traceme failed");
+    channel.close_read();
+    if (debug and ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
+      exit_with_perror(channel, "Traceme failed");
     }
     if (execlp(path.c_str(), path.c_str(), nullptr) < 0) {
-      error::send_errno("Execlp failed");
+      exit_with_perror(channel, "Execlp failed");
     }
   }
+
+  channel.close_write();
+  auto data = channel.read();
+  channel.close_read();
+
+  if (data.size() > 0) {
+    waitpid(pid, nullptr, 0);
+    auto chars = reinterpret_cast<char *>(data.data());
+    error::send(std::string(chars, chars + data.size()));
+  }
+
   std::unique_ptr<rdb::process> proc(
-      new rdb::process(pid, /*terminate_on_end=*/true));
-  proc->wait_on_signal();
+      new rdb::process(pid, /*terminate_on_end=*/true, debug));
+
+  if (debug) {
+    proc->wait_on_signal();
+  }
 
   return proc;
 }
@@ -41,7 +68,7 @@ std::unique_ptr<rdb::process> rdb::process::attach(pid_t pid) {
     error::send_errno("Couldn't attach to process");
   }
   std::unique_ptr<rdb::process> proc(
-      new rdb::process(pid, /*terminate_on_end=*/false));
+      new rdb::process(pid, /*terminate_on_end=*/false, /*is_attached=*/true));
   proc->wait_on_signal();
 
   return proc;
@@ -82,17 +109,18 @@ rdb::stop_reason::stop_reason(int wait_status) {
 rdb::process::~process() {
   if (pid_ > 0) {
     int status;
+    if (is_attached_) {
+      if (state_ == process_state::running) {
+        kill(pid_, SIGSTOP);
+        waitpid(pid_, &status, 0);
+      }
+      ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+      kill(pid_, SIGCONT);
 
-    if (state_ == process_state::running) {
-      kill(pid_, SIGSTOP);
-      waitpid(pid_, &status, 0);
-    }
-    ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
-    kill(pid_, SIGCONT);
-
-    if (terminate_on_end_) {
-      kill(pid_, SIGINT);
-      waitpid(pid_, &status, 0);
+      if (terminate_on_end_) {
+        kill(pid_, SIGINT);
+        waitpid(pid_, &status, 0);
+      }
     }
   }
 }
